+++ I2C_LCD.h
@@ -84,8 +84,8 @@ public:
 
   //  PRINT INTERFACE ++
   size_t    write(uint8_t c);
-  size_t    center(uint8_t row, char * message);
-  size_t    right(uint8_t col, uint8_t row, char * message);
+  size_t    center(uint8_t row, const char * message);
+  size_t    right(uint8_t col, uint8_t row, const char * message);
 
   //  DEBUG  development
   uint8_t   getColumn() { return _pos; };  //  works.